Add Tofino support for ChassisConfig
This commit does the following:
 * Introduce a BF_PAL wrapper. The wrapper implements a C++ interface
   which can be used to mock BF_PAL in unit tests.
 * Implement BFChassisManager::PushChassisConfig to add / configure /
   enable device ports. This code may change slightly as more "getters"
   are added to the BF_PAL wrapper.
 * Better support for retrieving port data (now includes AdminStatus,
   PortSpeed, NegotiatedPortSpeed and AutonegStatus, in addition to
   OperStatus and PortCounters).
 * Automatic port configuration replay when a FastRefresh operation
   (SetForwardingPipelineConfig RPC) is performed.

+++ stratum/hal/lib/bmv2/bmv2_chassis_manager.h
@@ -99,19 +99,19 @@ class Bmv2ChassisManager {
 
   // Forward PortStatus changed events through the appropriate node's registered
   // ChannelWriter<GnmiEventPtr> object.
-  void SendPortOperStateGnmiEvent(uint64 node_id, uint64 port_id,
+  void SendPortOperStateGnmiEvent(uint64 node_id, uint32 port_id,
                                   PortState new_state)
       LOCKS_EXCLUDED(gnmi_event_lock_);
 
   friend ::util::Status PortStatusChangeCb(Bmv2ChassisManager* chassis_mgr,
                                            uint64 node_id,
-                                           uint64 port_id,
+                                           uint32 port_id,
                                            PortState new_state)
       LOCKS_EXCLUDED(chassis_lock);
 
   ::util::StatusOr<const SingletonPort*> GetSingletonPort(
-       uint64 node_id, uint64 port_id) const
-        SHARED_LOCKS_REQUIRED(chassis_lock);
+      uint64 node_id, uint32 port_id) const
+      SHARED_LOCKS_REQUIRED(chassis_lock);
 
   bool initialized_ GUARDED_BY(chassis_lock);
